Organize remaining models into folders and split into extensions
