+++ App/Sources/Core/Models/Commands/Application/ApplicationCommand+Modifier.swift
