+++ App/Sources/Core/Models/Commands/Application/ApplicationCommand.swift
