+++ App/Sources/Core/Models/Commands/BuiltIn/BuiltInCommand+Kind.swift
