+++ App/Sources/Core/Models/Commands/Keyboard/KeyboardCommand+Kind.swift
