+++ App/Sources/Core/Models/Commands/MenuBar/MenuBarCommand+Token.swift
