+++ App/Sources/Core/Models/Commands/MenuBar/MenuBarCommand.swift
