+++ App/Sources/Core/Models/Commands/Open/OpenCommand.swift
