+++ App/Sources/Core/Models/Commands/Script/ScriptCommand+Kind.swift
