+++ App/Sources/Core/Models/Commands/Text/TextCommand.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+struct TextCommand: MetaDataProviding {
+  var kind: Kind
+
+  init(_ kind: Kind) {
+    self.kind = kind
+  }
+
+  func copy() -> TextCommand {
+    TextCommand(kind.copy())
+  }
+}
