+++ App/Sources/Core/Models/Commands/WindowFocus/WindowFocusCommand.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+struct WindowFocusCommand: MetaDataProviding {
+  var kind: Kind
+  var meta: Command.MetaData
+
+  func copy() -> WindowFocusCommand {
+    WindowFocusCommand(kind: kind, meta: meta.copy())
+  }
+}
